# Optimize: Switch to no-fork provider architecture