--- a/apis/mdb/v1alpha1/zz_mongodbuser_types.go
+++ b/apis/mdb/v1alpha1/zz_mongodbuser_types.go
@@ -22,9 +22,6 @@ type MongodbUserInitParameters struct {
 	// +kubebuilder:validation:Optional
 	ClusterIDSelector *v1.Selector `json:"clusterIdSelector,omitempty" tf:"-"`
 
-	// The name of the user.
-	Name *string `json:"name,omitempty" tf:"name,omitempty"`
-
 	// Set of permissions granted to the user. The structure is documented below.
 	Permission []MongodbUserPermissionInitParameters `json:"permission,omitempty" tf:"permission,omitempty"`
 }
@@ -33,9 +30,6 @@ type MongodbUserObservation struct {
 	ClusterID *string `json:"clusterId,omitempty" tf:"cluster_id,omitempty"`
 
 	ID *string `json:"id,omitempty" tf:"id,omitempty"`
-
-	// The name of the user.
-	Name *string `json:"name,omitempty" tf:"name,omitempty"`
 
 	// Set of permissions granted to the user. The structure is documented below.
 	Permission []MongodbUserPermissionObservation `json:"permission,omitempty" tf:"permission,omitempty"`
@@ -54,10 +48,6 @@ type MongodbUserParameters struct {
 	// Selector for a MongodbCluster to populate clusterId.
 	// +kubebuilder:validation:Optional
 	ClusterIDSelector *v1.Selector `json:"clusterIdSelector,omitempty" tf:"-"`
-
-	// The name of the user.
-	// +kubebuilder:validation:Optional
-	Name *string `json:"name,omitempty" tf:"name,omitempty"`
 
 	// The password of the user.
 	// +kubebuilder:validation:Optional
@@ -136,7 +126,6 @@ type MongodbUserStatus struct {
 type MongodbUser struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// +kubebuilder:validation:XValidation:rule="!('*' in self.managementPolicies || 'Create' in self.managementPolicies || 'Update' in self.managementPolicies) || has(self.forProvider.name) || (has(self.initProvider) && has(self.initProvider.name))",message="spec.forProvider.name is a required parameter"
 	// +kubebuilder:validation:XValidation:rule="!('*' in self.managementPolicies || 'Create' in self.managementPolicies || 'Update' in self.managementPolicies) || has(self.forProvider.passwordSecretRef)",message="spec.forProvider.passwordSecretRef is a required parameter"
 	Spec   MongodbUserSpec   `json:"spec"`
 	Status MongodbUserStatus `json:"status,omitempty"`