--- a/cluster/images/provider-upjet-yc/terraformrc.hcl
+++ b/cluster/images/provider-upjet-yc/terraformrc.hcl
@@ -1,9 +1,0 @@
-provider_installation {
-  filesystem_mirror {
-    path    = "/terraform/provider-mirror"
-    include = ["*/*"]
-  }
-  direct {
-    exclude = ["*/*"]
-  }
-}