… apis/compute/v1alpha1/zz_generated.deepcopy.go
… apis/mdb/v1alpha1/zz_generated.deepcopy.go
… apis/storage/v1alpha1/zz_bucket_types.go
… apis/storage/v1alpha1/zz_generated.resolvers.go
… apis/ydb/v1alpha1/zz_generated.deepcopy.go
… apis/ydb/v1alpha1/zz_generated.resolvers.go
… apis/ydb/v1alpha1/zz_table_types.go
… cluster/images/provider-upjet-yc/Dockerfile
… cluster/images/provider-upjet-yc/Makefile
… cmd/generator/main.go
… cmd/provider/main.go
… config/external_name.go
… config/iam/config.go
… config/provider.go
… examples-generated/mdb/v1alpha1/mongodbdatabase.yaml
… examples-generated/mdb/v1alpha1/mongodbuser.yaml
… examples-generated/mdb/v1alpha1/mysqluser.yaml
… examples-generated/mdb/v1alpha1/postgresqldatabase.yaml
… examples-generated/ydb/v1alpha1/table.yaml
… go.mod
… go.sum
… internal/clients/yandex-cloud.go
… internal/controller/alb/backendgroup/zz_controller.go
… internal/controller/alb/httprouter/zz_controller.go
… internal/controller/alb/loadbalancer/zz_controller.go
… internal/controller/alb/targetgroup/zz_controller.go
… internal/controller/alb/virtualhost/zz_controller.go
… internal/controller/audit/trailstrail/zz_controller.go
… internal/controller/compute/disk/zz_controller.go
… internal/controller/compute/diskplacementgroup/zz_controller.go