### Description of your changes
In this pull request, the terraform binary was abandoned by switching to a no-fork provider architecture. This speeded up and improved its performance

![image](https://github.com/tagesjump/provider-upjet-yc/assets/50084893/533e1fad-55cb-40dd-90db-fd83b4dd72c4)

|   Metric                    | Old Architecture | New Architecture | Improvement       |
|-----------------------------|------------------|------------------|-------------------|
| Average Time To Ready State |        19.55 min |        1.012 sec |     1159x speedup |
| Peak Time To Ready State    |        95.45 min |         4.00 sec |     1432x speedup |
| Average Memory Utilization  |         1.23 GiB |          405 MiB |     60% reduction |
| Peak Memory                 |         1.79 GiB |          421 MiB |     77% reduction |
| Average CPU                 |           95.77% |            2.91% | 92.86% difference |
| Peak CPU                    |           98.17% |            3.46% | 94.71% difference |

I have:

- [x] Read and followed Crossplane's [contribution process].
- [x] Run `make reviewable test` to ensure this PR is ready for review.